refactor: remove asset widget hook and merge it with cost  widget hook
